feature: abstract cloneable logic into CanBeCloned
